requirements.txt
accelerate
argh
dacite
demucs
diffusers
flask
flask_cors
numpy
pillow
plotly
pydub
pysoundfile
scipy
soundfile
sox
streamlit
torch
torchaudio
torchvision
transformers

